requirements.txt
flask>=2.0.0
requests>=2.26.0
python-dotenv>=0.19.0
pydantic>=1.9.0  # Для валидации (опционально)